feat: Vertical 카드 컴포넌트
