+++ src/assets/svgs/CartIcon.tsx
@@ -8,13 +8,13 @@ const SvgCartIcon = (props: SVGProps<SVGSVGElement>) => (
   >
     <path
       stroke="#374553"
-      stroke-width={1.2}
+      strokeWidth={1.2}
       d="M4.8002 5.59998H15.2002L16.8002 16.8H3.2002L4.8002 5.59998Z"
     />
     <path
       stroke="#374553"
-      stroke-width={1.2}
-      stroke-linecap="round"
+      strokeWidth={1.2}
+      strokeLinecap="round"
       d="M12 7.19998V3.59998C12 2.49541 11.1046 1.59998 10 1.59998V1.59998C8.89543 1.59998 8 2.4954 8 3.59997V7.19998"
     />
   </svg>
